+++ src/features/hooks/common/useThrottle.ts
@@ -1,12 +1,13 @@
-import { DebounceType } from '@devTypes/common/debounce.type';
 import { useCallback, useState } from 'react';
-import { debounce } from '@devUtils/common/debounce';
+import { ThrottleTypes } from '@dev/shared/types/common/throttle.type';
+import throttle from '@devUtils/common/throttle';
 
-export const useDebounce = ({ fn, wait = 500 }: DebounceType) => {
+export const useThrottle = ({ fn, wait = 500 }: ThrottleTypes) => {
     // 상태 관리 (선택 사항, 필요하다면 사용할 수 있음)
     const [state, setState] = useState(false);
+
     const event = useCallback(
-        debounce({
+        throttle({
             fn: (...args: unknown[]) => {
                 // 실제로 호출된 함수가 실행될 때 상태 변경
                 setState(true);
