+++ src/features/utils/common/debounce.ts
@@ -1,4 +1,4 @@
-import { DebounceType } from '@devTypes/common/debounce.type';
+import { DebounceType } from '@dev/shared/types/common/debounce.type';
 
 export const debounce = ({ fn, wait = 500 }: DebounceType) => {
     let debounceCall: ReturnType<typeof setTimeout> | null = null;
