+++ src/shared/assets/styles/index.css
