+++ src/shared/assets/styles/tailwind.css
